backend/requirements.txt
torchscale
pinecone
python-dotenv
flask
flask_cors
pyvi
transformers
sentence-transformers
pandas
huggingface_hub
gunicorn